Initial working prototype of exome and variant viewer.
Created a visualization of exons in a gene using SVG rectangles. Reduced introns (gap length between exon rectangles) to a uniform length. Exons are scaled proportionately but introns are not. Created lollipops below exon blocks to represent variants in the gene, mapped the variants to the correct position on the exon using d3.scaleLinear(), where the domain is the start position of each exon and intron given their original intron lengths, while the range is the start position of each exon and intron given uniform intron lengths. When user hovers over a lollipop circle, a tooltip appears that displays data for that variant.