Created function for calculating y axis scale
Function getYAxisScaleFromVariable() takes in variable name such as alleleNumber or 'MAF' and then finds the appropriate variable from the variant data. It generates a scale function that maps variable values to the chart height. This is then used by function addLollipopToChart() to place lollipop markers in the correct height according to the scale.