Created function that updates zoomed in chart based on yAxisVariable updates
Function redraws y axis scale, y axis label, and variant heights based on the updated y axis variable. Connecting this function to the buttons will be difficult though, and it needs to return handle to lollipop circles so that upon mouse hovering on the lollipop circles, the variant data appears.