Consolidated all the svg handles and svg transform handles into a single area in the code
By organizing all the transforms into one place, it will be easier to understand which transforms will be updated when the y axis variable changes. The variant transforms for the zoomed out and zoomed in charts, as well as the y axis variable transform for the zoomed in chart, will be updated.