Created svgTransform wrapper object
To update the y axis variables, I will need to update the contents of three transforms - variantTransformZoomedOut, variantTransformZoomedIn, yAxisTransformZoomedIn. The number of variables required to update three svg transforms is too much for one update function to take in. To reduce complexity, create a generic transform wrapper function that contains two inputs. 1) An svg transform handle and 2) a dictionary of arguments needed to render the data for a specific transform. This will force me to rewrite several of the chart generation functions that previously used several inputs and a transform handle.