Writing your first Django app, part 2
Renamed tables in db.sqlite3 file. Table 'exons' renamed to 'visualization_gene'. Table 'variants_SAMD11' renamed to 'visualization_variants_SAMD11'. This follows the Django db naming convention where the first word in the table name is the name of the application that will access the table. I am unsure if 'visualization_variants_SAMD11' can be mapped to the Variants model, or if I have to rename the table to 'visualization_variants'.