Writing your first Django app, part 2
Ran migrations for the Gene class and the Variant class in exomesite/visualization/models.py. Edited four columns in the Variant class to allow blanks - consequence, proteinconsequence, transcriptconsequence, and flags. Re-imported the csv files containing exon data and variant data into the tables 'visualization_gene' and 'visualization_variant' generated by the migration command. Had to add an auto-increment column in the variants.csv file because the model specifies an autoincrememnt field called 'id' that serves as the primary key of the table - otherwise the sqlite import command would not work.