Writing your first Django app, part 3
Added results view in exomesite/visualization/views.py which is linked to exomesite/visualization/templates/visualization/results.html. This displays the details of the genes exons using d3 js library. If there are multiple variants of the same gene in the visualization_gene table, then the variation of the gene with the most exons is displayed.