Writing your first Django app, part 4
Modified templates/visualization/index.html to display a searchbox and a search button for querying based on gene names. Modified exomesite/visualization/views.py to support a new search view, which takes the string entered from the searchbox in index.html and redirects that string to the result view so that a query based on the gene name can be executed. Added the 'search' path to urls.py; because result takes the pattern 'visualization/<string>', it mistook 'visualization/search' to be a query where gene_name = 'search'. Therefore, I had to place the 'search' path above the 'result' path in urls.py.