Writing your first Django app, part 4
Modified the result view in views.py to redirect the user to index view if the gene name in the searchbox does not match any entries in the visualzation_gene table. Created a function get_ten_genes() in views.py that gets up to the first 10 uniquely named genes from the visualization_gene table. This modularizes code used in the index view and the result view.