Writing your first Django app, part 6
Skipped ahead of part 5 of the Django 2.0 tutorial which is about writing tests.Went to chapter 6 which is about creating a static folder in the visualization app folder. Static folder stores the css and js files for the visualization app. Created static/visualization/css/style.css file and as a initial test, turned the sample list of genes blue and the error message upon searching a blank searchbox with no gene name inputted into a red colored text.