Integrating development-mini branch with development-django branch
Modified templates/visualization/result.html to have the same html body as index.html in the development-mini branch. The goal is to integrate the front end code in index.html with the backend code that I have written in the django-development branch by slowly taking code from the development-mini branch and building it up in result.html.