Added function to limit the length of non coding parts of the first exon and last exon based on input variable nonCodingLimitLength
Modified static/visualization/js/gene.js to contain function limitNonCodingExonLength(), which limits the length of the non-coding part of the first exon and last exon to nonCodingLimitLength, which is currently set to 200. So if cdsStart = 5000 and exonStarts[0].start = 4000, then the non-coding part of the first exon is from the base pair at position 4000 to the base pair at position 5000. This length is 5000 - 4000 = 1000, which exceeds the limitNonCodingExonLength, so exonStarts[0].start is set to 4800. Now 5000 - 4800 = 200, which is within the limit of limitNonCodingExonLength.