Added function to calculate uniform intron lengths as a ratio of the sum of the exon lengths
The actual intron lengths between exons in a gene can often be much larger than the exon lengths. If we render introns by their actual lengths, they would crowd out exons in the d3.js charts. Therefore, in the chart I use a uniform intron length that guarantees that the sum of the introns will take up half the length of the sum of the exons, and thus allow the exons to be displayed prominently in the chart. Later I will create scaling functions that map a position from its actual intron length position to its uniform intron length position.