Added functions to calculate values of exonStarts and exonEnds after subtracting by an offset value
The base pair position of exonStarts and exonEnds can range in the millions so we have to subtract by an offset value equal to exonStarts (after exonStarts is adjusted by limitNonCodingExonLengths(), which limits the non-coding parts of the first and last exons to a length of 200). This way, the start position of the first exon will be zero, and it will be easy to put these values into scale functions for a d3.js chart.