Modified functions that get exonStarts array and exonEnds array
Instead of storing two int arrays, exonStarts and exonEnds, the Gene object stores an array of Exon objects, each of which stores a start and end. The functions getExonStarts() and getExonEnds() are used to derive exonStarts and exonEnds from the array of Exon objects. This allows a clear object oriented hierarchy.