Modified functions for processing exon data to focus on creating arrays of exon objects first, and then deriving exonStarts and exonEnds through them
Modified static/visualization/js/gene.js so that when calculating exonStarts andexonEnds assuming a uniform intron length or being subtracted by the offset value of the starting position of the first exon, we first create an array of exons containing those newly calculated values. Then the exonStarts and exonEnds functions can be used to retrieve the new values.