Modified getExonStarts() and getExonEnds() to take in an array of Exon objects
The arrays exonStarts and exonEnds have to be modified throughout result.html into two formats. One format is the original values of exonStarts and exonEnds, minus an offset equal to exonStarts[0] after the non-coding part of the first exon is limited to a max length of 200 base pairs. The second format is where each value in exonStarts and exonEnds is modified assuming a uniform intron length, where the sum of those uniform intron lengths is equal to half the length of the sum of the exons. This second format is also subtracted by the offset of exonStarts[0]. For each modification, create a function that returns an array of Exons with modified start and end positions, and then pass that array of Exons into getExonStarts() and getExonEnds() to get array values needed by scale functions that map base pair positions from the original intron lengths to the uniform intron lengths.