Added function to calculate the length of introns
Calculating the length of introns will help verify that scaling functions mapping base pair positions from original intron lengths to uniform intron lengths are working properly. This is because depending on the length of the intron, I will have to split the intron into three parts on the scale function mapping original to uniform. The two parts on the side of the intron contain the most variants, while the part in the middle does not contain many, if any at all, variants. Therefore, in the d3 js chart, the three parts will be given equal lengths on the chart so that the middle part lacking variant data will not dominate the viewing space.