Added function to calculate range values that are mapped to the domain values returned from getDomain()
Modified static/visualization/js/gene.js by adding a function getRange() which takes in basePairsOutsideExonLimit, the limit on the number of base pairs outside of an exon that will be included as a partition in an intron. Currently, basePiarsOutsideExonLimit = 50, which means that if a intron spans from base pair position 4000 to 5000, then it will be divided into three parts by the getDomain() function: 4000-4050, 4050-4950, 4950-5000. The getRange() function calculates the corresponding base pair position values given uniform intron lengths. So if the uniform intron length = 150, then the positions from getDomain() would be mapped to: 4000 to 4050, 4050-4100,4100-4150. The exon following this intron has an original starting position of 5000 that would be mapped to a new position of 4150 given the uniform intron length modifier.