Added scale functions that use the scaleLinear() function in d3 js to map the original base pair positions to positions in the chart.
In static/visualization/js/chart.js, I added three new functions. Function getScaleOriginalIntronsToUniformIntrons() maps the base pair positions of exons assuming the original intron lengths to new positions in the uniform intron length. Function getScaleUniformIntronsToChart() maps the new positions in the uniform intron length to positions on the d3 svg chart. Function addRectToTransform() takes a svgTransform handle as an input and draws rectangles representing exons onto the transform handle. It uses the two aforementioned scale functions to draw the start position and the lengths of the rectangles correctly. All functions are part of the ChartController class.