Added function getIntronsWithUniformIntronLength(), which recalculates the start  and end positions of introns assuming a uniform intron length
Modified static/visualization/js/gene.js by adding function getIntronsWithUniformIntronLength() to Gene class. It uses the getScaleOriginalIntronsToUniformIntrons() function to calculate the new positions of the introns, thereby not requiring the extra additions involved in getExonsWithUniformIntronLengths(). Introns with uniform intron length will later be split into three parts, scaled to fit into the chart, and displayed as dashed lines. I also moved the getScaleOriginalIntronsToUniformIntrons() function from the ChartController class in chart.js to the Gene class in gene.js because the processed data required by that scale function belongs in the Gene class. Meanwhile, I kept getScaleUniformIntronsToChart() in the ChartController class because that function is more about presenting data in a viewable format, instead of calculating exons and introns.