Added function addLinesToZoomedOutChart() which takes a handle to an svg transform and draws dashed lines representing introns into that transform.
Modified static/visualization/js/chart.js by adding the function addLinesToZoomedOutChart(), which draws dashed lines representing introns into an input svg transform handle. The introns have modified starting and ending positions according to the uniform intron length, and are then split into partitions of three, and then turned into a data array to be inputted to the d3 path drawing function. The data array consists of an array of x,y coordinates, where each coordinate is itself an array of two numbers. The middle partition of each intron is drawn with a light grey and smaller dashes, while the left and right partitions of each intron is drawn with a black color and bigger dashes. This emphasis of the side partitions of the intron reflects the fact that most variants from the gene datasets are contained within the left and right partitions, which are 50 base pairs outside of the nearest exon.