Modified function initializeExons in the Gene class to exclude exons that end before cdsStart or exons that start after cdsEnd.
I noticed an error in the display of the gene SAMD11(NM_15486.2) where the first exon and the second exon both were displayed in a shorter height representing non-coding areas. Upon inspection, I found that the starting and ending positions of the first exon in SAMD11 both occured before cdsStart. So I modified initializeExons() to exclude exons in this exceptional case. Then I removed the exonCount variable from the Gene class and replaced it with a getExonCount() function that returns exons.length. This way, if the original exonCount passed from the json object in result.html contains an exon that is excluded, the exon array will become shorter and getExonCount() will return a value that accurately captures this modification of the exon array.