Added initializeVariants() to Gene class, which takes a list of variants and filters out those that exceed the boundaries of the Gene object.
Boundaries of the Gene object here refer to the start position of the first exon and the end position of the last exon, after calling initializeExons() and limitNonCodingExonLength(). The function initializeExons() filters out exons that end before cdsStart and exons that start after cdsEnd. The limitNonCodingExonLength() function limits the length of non-protein-coding areas of an Exon object to be at most of length = 200. After these adjustments to the exons, the filtering of variants occurs in initializeExons() and the resulting array of variants is set to be the variant variable in the Gene object.