Added prototype of function addVariantsToTransform(), which draws lollipops on the chart that represent positions of variants in alleles within the gene
Modified static/visualization/js/chart.js by adding function addVariantsToTransform(). This a prototype of the function and is incomplete because it does not have color codings for different types of variants, the option to choose between different y axis variables and scale the height of the lollipop markers based on that y axis variable, or the ability to hover over a marker and seeing a tooltip display data on that variant. These are all features that will be gradually built.