Deleted migrations and re-ran them because of an error in the database where every field after the id column had a value that was shifted one column to the right.
Deleted the visualization_gene and visualization_variant tables from the db.sqlite3 file. Then I re-ran migrations for the Gene and Variant classes in exomesite/visualization/models.py. Then I imported the gene and variant csv files (both files contain a column id which starts from 1 and increments by 1 for each row). Checked the visualization_gene and visualization_variant tables, and the fields align with the json object that is sent to exomesite/visualization/templates/result.html. When parsed, the json object containing the list of variants has the correct values for the flags and annotation fields.