Added functions to calculate the width and height of the chart after subtracting by the margins, and getYAxisScale(), which returns scales the values of a y axis variable such as allele number or allele frequency, to the height of the chart after accounting for margins
Modified static/visualization/js/chart.js by adding functions getChartHeight() and getChartWidth(), which return the values of the height and width of the chart after accounting for margins. Added function yAxisScale which currently can take two y axis variable names: 'MAF' for allelefrequency and 'alleleNumber'. It returns a scale function that maps each of values for a variable name to its corresponding y position in the chart.