Modified addVariantsToChart() function to display a tooltip that shows data on a variant when hovering over its lollipop representation on the chart.
When hovering over the circle of a lollipop, the tooltip displays data on the variant in the position. Currently I have it set so that if a base pair position contains multiple variants, only the data on the first variant is shown. Later it would be good to figure out a way to display data on all of those variants. Also modified static/visualization/css/style.css to make the tooltip to appear next to where the user is hovering and display text nicely.