Modified Gene class to take basePairsOutSideExonLimit as an input and replace functions that required basePairsOutSideExonLimit as an input with gene.getBasePairOutsideExonLimit() function
Noticed that the variants within intron regions were not being placed in the correct x positions on the chart. The problem was that getScaleOriginalIntronsToUniformIntrons() required basePairsOutsideExonLimit as an input. Instead of passing that input to the function, I decided to make basePairsOutsideExonLimit a variable in the Gene class and use a getter function. Error was that I was passing the parsed json object containing variant data instead of the basePairsOutsideExonLimit.