Fixed bug caused when changing the y axis variable being displayed, the tooltip would no longer appear when hovering over the lollipop markers representing variants
Instead of keeping the tooltip as a global variable in result.html, I moved it inside of the ChartController class. Then, in the function addVariantsToTransform(), I had it get the tooltip by calling getTooltip() instead of passing in the tooltip from the drawZoomedOutChart() function. This fixed the scope of the tooltip, which was lost originally. I also changed the name of drawZoomedInChart() tohighlightHoveredExon() and added a variable called hoveredExonIndex to the ChartController class. That way, whenever highlightHoveredExon() detects the user highlighting over a exon rectangle, it only needs to call setHoveredExonIndex(). Later on, I will implement the drawZoomedInChart() function, which will use this new hovered exon index to display the proper exon in detail.