Moved the limitNonCodingExonLength() function to be called in the Gene class constructor instead of in the result.html
Removed dependency on outside objects or scripts to call the limitNonCodingExonLength() function in the Gene class. By calling it in the constructor, the Gene class contains and initializes all data processing of the exon and variant data. The chart class can simply call all this data without having to call extra processes.