Added function addZoomedInExonToTransform(), which draws a zoomed in view of the exon that is hovered over in the zoomed out view chart.
Modified static/visualization/js/chart.js by adding addZoomedInExonToTransform(), which calculates the scale function needed to display a zoomed in view of a single exon that is currently hovered over in the zoomed out view chart. The function is called by drawZoomedInChart(), which is in turn called by highlightHoveredExon(), which is in turn called by addZoomedOutExonsToTransform(), which is in turn called by drawZoomedOutChart().