Fixed bug that caused the first and second exons in PCSK9 and other genes to be clumped together with the intron between them in the chart.
Bug was located in the getExonsWithUniformIntronLength() function in static/visualization/js/gene.js. The problem was that at the end of the function, initializeExons() was being called to initialize an Exon array with uniform intron length. But I had already changed initializeExons() to include the trimming of exon non coding partitions by calling limitNonCodingLength(). Somehow the first and second exons with uniform intron length got trimmed, and the resulting exonsWithUniformIntronLength array returned to the getRange() function then messed up the scaleOriginalIntronsToUniformIntrons() function.