Added function addZoomedInIntronsToTransform(), which draws the introns for the hovered over exon in the zoomed in chart view.
The function addZoomedInIntronsToTransform takes the svg transform that holds exons in the zoomed in chart view, and draws dashed lines representing introns to  the sides of the exons. It uses the hoveredExon input to calculate the scale function needed to place the intron lines correctly at the sides of the exon while maintaining the margins on the chart. I realized that both addZoomedInIntronsToTransform and addZoomedInExonsToTransform would benefit from getting hoveredExon as an input as it would reduce the number of redundant subroutin calls on getting the hovered index and the hovered over exon.