Added function addZoomedInNonCodingExonPartitionsToTransform(), which draws white rectangles on the non-coding parts of a hovered over exon to distinguish the height of the protein-coding and non-protein-coding parts.
When an exon is hovered over, drawZoomedInChart() calls addZoomedInNonCodingExonPartitionsToTransform(), which in turns calls getNonCodingExonIndices() from the Gene class to see if the current hovered over exon contains a non-protein-coding part or not. If so, then white rectangles are drawn on the top and bottom of the colored rectangle representing in hovered over exon in the zoomed in chart view.