Moved code for initializing data and charts from result.html to main.js file
Created file main.js to store code for initializing data processing and drawing charts that used to be stored in result.html. This makes the code more modular. Only the global variable chartController and a call the initializeCharts() in main.js remains in the script tag of the result.html file. Changed views.py to return a Queryset of genes that match the searched for a gene name. Originally I filtered out the gene object with the most exons in views.py, but this filtering prevented the Gene objects from being json serialized. By returning the Gene objects as a Queryset that is json-serializable, I can get all the attributes from the gene table. In utils.py, the function getGeneTranscriptWithMostExons() takes the json object containing the Gene Queryset and returns the gene object with the highest exoncount.